Use uint64_t for C++ implementation

+++ day22/experiments/day22.cpp
@@ -1,12 +1,13 @@
 #include <cstdlib>
+#include <cstdint>
 #include <iostream>
 #include <fstream>
 #include <vector>
 
-const int LIMIT = 2000;
+const uint64_t LIMIT = 2000;
 
-int next(int secret) {
-  int mask = 16777215; // = 2^24 - 1
+uint64_t next(uint64_t secret) {
+  uint64_t mask = 16777215; // = 2^24 - 1
   secret ^= secret << 6;
   secret &= mask;
   secret ^= secret >> 5;
@@ -16,19 +17,19 @@ int next(int secret) {
   return secret;
 }
 
-int prng(int secret, int n) {
-  for (int i = 0; i < n; i++) {
+uint64_t prng(uint64_t secret, uint64_t n) {
+  for (uint64_t i = 0; i < n; i++) {
     secret = next(secret);
   }
   return secret;
 }
 
-int monkey(int secret, int x1, int x2, int x3, int x4) {
-  int d1 = -1, d2 = -1, d3 = -1, d4 = -1;
-  for (int i = 0; i < LIMIT; i++) {
-    int lastPrice = secret % 10;
+uint64_t monkey(uint64_t secret, uint64_t x1, uint64_t x2, uint64_t x3, uint64_t x4) {
+  uint64_t d1 = -1, d2 = -1, d3 = -1, d4 = -1;
+  for (uint64_t i = 0; i < LIMIT; i++) {
+    uint64_t lastPrice = secret % 10;
     secret = next(secret);
-    int price = secret % 10;
+    uint64_t price = secret % 10;
     d1 = d2;
     d2 = d3;
     d3 = d4;
@@ -40,10 +41,10 @@ int monkey(int secret, int x1, int x2, int x3, int x4) {
   return -1;
 }
 
-int score(std::vector<int> input, int x1, int x2, int x3, int x4) {
-  int sum = 0;
-  for (int n : input) {
-    int price = monkey(n, x1, x2, x3, x4);
+uint64_t score(std::vector<uint64_t> input, uint64_t x1, uint64_t x2, uint64_t x3, uint64_t x4) {
+  uint64_t sum = 0;
+  for (uint64_t n : input) {
+    uint64_t price = monkey(n, x1, x2, x3, x4);
     if (price > 0) {
       sum += price;
     }
@@ -51,16 +52,16 @@ int score(std::vector<int> input, int x1, int x2, int x3, int x4) {
   return sum;
 }
 
-int findBestScore(std::vector<int> input) {
-  int bestScore = 0;
-  int bound = 9;
+uint64_t findBestScore(std::vector<uint64_t> input) {
+  uint64_t bestScore = 0;
+  uint64_t bound = 9;
   #pragma omp parallel for
-  for (int x1 = -bound; x1 <= bound; x1++) {
-    for (int x2 = -bound; x2 <= bound; x2++) {
+  for (uint64_t x1 = -bound; x1 <= bound; x1++) {
+    for (uint64_t x2 = -bound; x2 <= bound; x2++) {
       std::cout << "Searching (" << x1 << ", " << x2 << ")" << std::endl;
-      for (int x3 = -bound; x3 <= bound; x3++) {
-        for (int x4 = -bound; x4 <= bound; x4++) {
-          int n = score(input, x1, x2, x3, x4);
+      for (uint64_t x3 = -bound; x3 <= bound; x3++) {
+        for (uint64_t x4 = -bound; x4 <= bound; x4++) {
+          uint64_t n = score(input, x1, x2, x3, x4);
           if (n > bestScore) {
             bestScore = n;
           }
@@ -77,7 +78,7 @@ int main(int argc, char *argv[]) {
     return 1;
   }
 
-  std::vector<int> input;
+  std::vector<uint64_t> input;
 
   {
     std::ifstream file;
@@ -87,13 +88,13 @@ int main(int argc, char *argv[]) {
     }
   }
 
-  int part1 = 0;
-  for (int n : input) {
+  uint64_t part1 = 0;
+  for (uint64_t n : input) {
     part1 += prng(n, LIMIT);
   }
   std::cout << "Part 1: " << part1 << std::endl;
 
-  int part2 = findBestScore(input);
+  uint64_t part2 = findBestScore(input);
   std::cout << "Part 2: " << part2 << std::endl;
 
   return 0;
